+++ src/main/java/org/vcell/N5/UI/RemoteFileSelection.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="org.vcell.vcellfiji.UI.RemoteFileSelection">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="org.vcell.N5.UI.RemoteFileSelection">
   <grid id="27dc6" binding="mainPanel" layout-manager="FormLayout">
     <rowspec value="center:d:noGrow"/>
     <rowspec value="top:4dlu:noGrow"/>
